pyproject.toml
[project]
name = "agentic-code-review"
version = "1.0.0"
description = "Multi-agent AI code review bot powered by Groq + LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langchain-core>=1.2.17",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.10",
    "pydantic>=2.12.5",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "tenacity>=9.0.0",
    "uvicorn>=0.41.0",
]

[project.scripts]
code-review = "code_review.cli:main"
